Regras de validação no update

+++ app_locadora/app/Http/Controllers/MarcaController.php
@@ -16,13 +16,13 @@ class MarcaController extends Controller
     /**
      * Display a listing of the resource.
      *
-     * @return \Illuminate\Http\Response
+     * @return Marca[]|\Illuminate\Database\Eloquent\Collection|\Illuminate\Http\JsonResponse|\Illuminate\Http\Response
      */
     public function index() //get all
     {
         //
         $marcas = $this->marca->all();
-        return $marcas;
+        return  response()->json($marcas, 200);
     }
 
 
@@ -31,13 +31,16 @@ class MarcaController extends Controller
      * Store a newly created resource in storage.
      *
      * @param  \Illuminate\Http\Request  $request
-     * @return \Illuminate\Http\Response
+     * @return \Illuminate\Http\JsonResponse|\Illuminate\Http\Response
      */
     public function store(Request $request)
     {
         //
+
+        $request->validate($this->marca->rules(),$this->marca->feedbacks());
+
         $marcas = $this->marca->create($request->all());
-       return $marcas;
+       return response()->json($marcas, 201);
 
     }
 
@@ -45,13 +48,16 @@ class MarcaController extends Controller
      * Display the specified resource.
      *
      * @param  Integer
-     * @return \Illuminate\Http\Response
+     * @return \Illuminate\Http\JsonResponse|\Illuminate\Http\Response
      */
     public function show( $id) //get by id
     {
         //
         $marca = $this->marca->find($id);
-        return $marca;
+        if($marca === null){
+            return response()->json(['error'=>'Recurso pesquisado não existe'], 404) ;
+        }
+        return response()->json($marca, 200);
     }
 
 
@@ -60,27 +66,36 @@ class MarcaController extends Controller
      *
      * @param  \Illuminate\Http\Request  $request
      * @param  Integer
-     * @return \Illuminate\Http\Response
+     * @return \Illuminate\Http\JsonResponse|\Illuminate\Http\Response|string[]
      */
     public function update(Request $request, $id)
     {
-        //
         $marca = $this->marca->find($id);
+
+        if($marca === null) {
+            return response()->json(['erro' => 'Impossível realizar a atualização. O recurso solicitado não existe'], 404);
+        }
+
+        $request->validate($marca->rules(), $marca->feedback());
         $marca->update($request->all());
-        return $marca;
+        return response()->json($marca, 200);
+
     }
 
     /**
      * Remove the specified resource from storage.
      *
      * @param  Integer
-     * @return \Illuminate\Http\Response
+     * @return \Illuminate\Http\JsonResponse|\Illuminate\Http\Response|string[]
      */
     public function destroy($id)
     {
         //
         $marca = $this->marca->find($id);
+        if($marca === null){
+            return response()->json(['error'=>'Recurso pesquisado não existe'], 404) ;
+        }
         $marca->delete();
-        return ['msg'=>'A marca foi removida com sucesso'];
+        return response()->json(['msg'=>'A marca foi removida com sucesso'], 200);
     }
 }
